aider: Add a `cd` command to change the current working directory by updating the context data.
# Aider chat conversation:

USER: please add a cd() command that will change the current dir by updating the context data

ASSISTANT: ```diff
--- plugins/ah_files/mod.py
+++ plugins/ah_files/mod.py
@@ ... @@
+@command(is_local=True)
+async def cd(directory, context=None):
+    """Change the current working directory.
+
+    Parameter: directory - The directory to change to. This can be a relative or absolute path.
+
+    Example:
+
+    { "cd": "subdir1" }
+
+    Other Example (parent dir)
+
+    { "cd": ".." }
+
+    """
+    if 'current_dir' not in context.data:
+        context.data['current_dir'] = os.getcwd()
+    new_dir = os.path.realpath(os.path.join(context.data['current_dir'], directory))
+    if os.path.isdir(new_dir):
+        context.data['current_dir'] = new_dir
+        print(f'Changed current directory to {new_dir}')
+    else:
+        print(f'Directory {new_dir} does not exist.')
```
This diff adds a new command `cd` to the `mod.py` file, which changes the current working directory by updating the `context.data['current_dir']`. The command takes a `directory` parameter that can be a relative or absolute path. If the directory exists, it updates the context; otherwise, it prints an error message.