aider: Update ChatContext to save and load the persona name.
# Aider chat conversation:

USER: please update ChatContext so that it also saves and loads the persona name

ASSISTANT: To update `ChatContext` so that it also saves and loads the persona name, we need to modify the `save_context` and `load_context` methods in `plugins/ah_chat/chat.py`.

Here are the *SEARCH/REPLACE* blocks:

plugins/ah_chat/chat.py
```python
<<<<<<< SEARCH
        context_data = {
            'data': self.data,
            'chat_log': self.chat_log._get_log_data()
        }
=======
        context_data = {
            'data': self.data,
            'chat_log': self.chat_log._get_log_data(),
            'persona': self.persona
        }
>>>>>>> REPLACE
```

plugins/ah_chat/chat.py
```python
<<<<<<< SEARCH
        else:
=======
            self.persona = context_data.get('persona')
>>>>>>> REPLACE
```

This will ensure that the persona name is saved and loaded correctly in the `ChatContext`.