aider: Increased the access token expiration time from 30 minutes to 1 week.
# Aider chat conversation:

USER: great thanks. and if I do a fetch from chat.js will it automatically include the cookie? so I wouldn't actually need the authorization bearer header anymore? don't change any code yet please

ASSISTANT: No changes are needed to the files. The `fetch` API does not automatically include cookies in requests. However, you can configure the `credentials` option in the `fetch` options to include cookies. By default, the `credentials` option is set to `'same-origin'`, which means that cookies will be included for requests to the same origin as the website.

If you want to include cookies for all requests, regardless of the origin, you can set the `credentials` option to `'include'`. However, this is generally not recommended for security reasons, as it can expose sensitive information (such as authentication tokens) to third-party websites.

If you want to include cookies only for requests to specific origins, you can set the `credentials` option to an instance of the `RequestCredentials` object, which allows you to specify a list of allowed origins.

USER: no I meant if I make a request to the same domain, then those fetch requests include the cookie we set when we logged in?

ASSISTANT: No changes are needed.